Rename customizers
