+++ opentracing-spring-cloud-starter-jaeger/src/test/java/io/opentracing/contrib/spring/cloud/starter/jaeger/customizer/JaegerTracerB3CustomerizerEnabledSpringTest.java
@@ -17,8 +17,8 @@ package io.opentracing.contrib.spring.cloud.starter.jaeger.customizer;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.opentracing.contrib.spring.cloud.starter.jaeger.AbstractTracerSpringTest;
-import io.opentracing.contrib.spring.cloud.starter.jaeger.JaegerTracerCustomizer;
-import io.opentracing.contrib.spring.cloud.starter.jaeger.customizers.B3CodecJaegerTracerCustomizer;
+import io.opentracing.contrib.spring.cloud.starter.jaeger.TracerBuilderCustomizer;
+import io.opentracing.contrib.spring.cloud.starter.jaeger.customizers.B3CodecTracerBuilderCustomizer;
 import java.util.List;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,12 +33,12 @@ import org.springframework.test.context.TestPropertySource;
 public class JaegerTracerB3CustomerizerEnabledSpringTest extends AbstractTracerSpringTest {
 
   @Autowired
-  private List<JaegerTracerCustomizer> customizers;
+  private List<TracerBuilderCustomizer> customizers;
 
   @Test
   public void testCustomizersShouldContainB3Customizer() {
     assertThat(customizers)
         .isNotEmpty()
-        .extracting("class").contains(B3CodecJaegerTracerCustomizer.class);
+        .extracting("class").contains(B3CodecTracerBuilderCustomizer.class);
   }
 }
